Dear Valentin,
I am writing this message to point out to your attention some improvements that have been recently incorporated inside our brushing and linking utilitities.

Now it is possible to show the label of the rows (and not just the row numbers) both in the plot of brushed trajectories and in the associated scatter plot matrix.
See for example the screenshot below.
 

The code is given below for the famous car data example
   %%    Example of use of databrush with option 'label'
    % If option label is supplied (that is if rownames of the input matrix
    % are supplied), brushed units are shown with their names rather than
    % with their row numbers, both in the malfwdplot and in the associated
    % scatter plot matrix.
    close all
    load carsmall
    x1 = Weight;
    x2 = Horsepower;    % Contains NaN data
    y = MPG;    % response
    Y=[x1 x2 y];
    NameY={'Weight', 'HorsePower' 'MPG'};
   % Remove Nans
    boo=~isnan(y);
    Y=Y(boo,:);
    %   RowLabelsMatrixY is the cell which contains the names of the rows.
    RowLabelsMatrixY=cellstr(Model(boo,:));
    [fre]=unibiv(Y);
    m0=20;
    bs=fre(1:m0,1);
    [out]=FSMeda(Y,bs,'init',30);
    databrush=struct;
    % Write labels of trajectories inside the malfwdplot while brushing
    databrush.Label='on'; 
    % Do not remove labels after selection in the malfwdplot
    databrush.RemoveLabels='off';
    % Add the labels of the units in the associated scatter plot matrix
    databrush.labeladd='1'; % 
    % Just write labels for units which have a trajectory
    % greater than fground.fthresh.
    fground=struct;
    fground.fthresh=15;
    malfwdplot(out,'databrush',databrush,'label',RowLabelsMatrixY,'fground',fground,'nameY',NameY)
    
The second improvement concerns the persist option of databrush.
Now the (row)labels of the selected units after each brushing operation appear with a particular color (of course the same color which is used for the associated brushed points in the scatter plot matrix.

For example if in the example above you add 
databrush.persist=’on’

here is for example  what you get
 

Note that the labels after each brushing action remain in the malfwdplot (given that)
    % Write labels of trajectories inside the malfwdplot while brushing
    databrush.Label='on'; 
    % Do not remove labels after selection in the malfwdplot
    databrush.RemoveLabels='off';

while in the associated scatter plot matrix just the labels of the last set of brushed units remain (give that)

    % Add the labels of the units in the associated scatter plot matrix
    databrush.labeladd='1';


Now I am wondering whether we shoud have an option of the kind
databrush.labeladd='persist'

in order to leave the labels in the associated scatter plot matrix (???).

Finally: an important point:
Initially we used the names malfwdplot.m and resfwdplot., because we had in mind to monitor Mahalanobis distances (multivariate analysis) and residuals (regression).

The function malfwdplot however is very general and has been enhanced in such a way that it can cope not only with Mahalanobis distance but with any other estimator monitored (for a value of breakdown point, efficiency …).

For example, Silvia Salini in Cc had a sort of monitoring of the Gower index (proximity index) for mixed data and she wanted to show the brushing of the associated trajectories of the particular statistic she is monitoring in the space of the first 3 principal components (therefore not necessarily in the associated scatter plot matrix).

The only one thing she had to do was:
Create structure out as follows
out=struct;
SPHI2 is the matrix which contained the monitoring of the proximity Phi index 
out.MAL=SPHI2;
Y(:,1:3) is the matrix of the first 3 principal components (but it can be any matrix, e.g. even the matrix of robust principal components …)
out.Y=Y(:,1:3);
Of couse the labels of the columns of matrix Y must be suppled
% Labels of the columns of Y
NameY={'PC1' 'PC2' 'PC3'};

Applying the same syntax of the example above (in Silvia’s example the units are the countries) here is what you get.

 

This is simply to say that malfwdplot and resfwdplot are not only confined to pure multivariate analysis and regression, but can be easily adapted for any research who wants to explore brushing and linking monitoring of an estimator for a particular value of an input parameter (i.e. bdp, efficiency, restriction factor for tclust ….)

Now let me conclude my message with a request.

I am willing to offer a full dinner to the first person who will explain me how to make latex commands work inside the datatooltip.
 

There is the latex interpreter but I could not make it work. So, a dinner is waiting! All those who have received this message can participate

Best wishes and see you soon

Marco

p.s. Fabrizio who is the our “representative R user” will soon let you have comments on fsdaR


git-svn-id: https://160.78.46.112/svn/fsda.code.000/FSDA/trunk@1961 336e6153-a435-9040-aa94-c6aba0fb517b

Former-commit-id: 1a155ce8d482b3124bd2cf4726f7370ae11a04ae